==> book/04-data-wrangling.Rmd <==

# Data wrangling part one

```{r, child='_setup.Rmd'}
```

```{r, echo = F, warning = F, message = F}
library(tidyverse)
library(janitor)
library(lubridate)
library(here)
library(kableExtra)

penguins <- read_csv(here("book", "files", "penguins_raw.csv"))
```

It may surprise you to learn that scientists actually spend far more time cleaning and preparing their data than they spend actually analysing it. This means completing tasks such as cleaning up bad values, changing the structure of dataframes, reducing the data down to a subset of observations, and producing data summaries. 

Many people seem to operate under the assumption that the only option for data cleaning is the painstaking and time-consuming cutting and pasting of data within a spreadsheet program like Excel. We have witnessed students and colleagues waste days, weeks, and even months manually transforming their data in Excel, cutting, copying, and pasting data. Fixing up your data by hand is not only a terrible use of your time, but it is error-prone and not reproducible. Additionally, in this age where we can easily collect massive datasets online, you will not be able to organise, clean, and prepare these by hand.

In short, you will not thrive as a scientist if you do not learn some key data wrangling skills. Although every dataset presents unique challenges, there are some systematic principles you should follow that will make your analyses easier, less error-prone, more efficient, and more reproducible.

In this chapter you will see how data science skills will allow you to efficiently get answers to nearly any question you might want to ask about your data. By learning how to properly make your computer do the hard and boring work for you, you can focus on the bigger issues.

## Tidyverse

In the first chapter, we introduced a package called tidyverse, which forms the foundation of many data skills you will develop. The tidyverse, developed by a team led by the renowned data scientist Hadley Wickham (Wickham, 2017), is a collection of packages. Its core packages include dplyr, tidyr, readr, purrr, ggplot2, and tibble. These packages provide a comprehensive set of functions that cover data wrangling and visualization tasks.

When you load library(tidyverse) in your code, all these packages are loaded simultaneously. While you'll learn to use various functions from tidyverse progressively, this chapter focuses on dplyr for data wrangling and ggplot2 for visualization—creating images and figures.

Specifically, dplyr offers six essential functions, referred to as the "Wickham Six" or "one-table verbs." These functions are named based on common English verbs, making code more readable:

| Function  | Description                                            |
|-----------|--------------------------------------------------------|
| `select()`  | Include or exclude certain variables (columns)          |
| `filter()`  | Include or exclude certain observations (rows)          |
| `mutate()`  | Create new variables (columns)                          |
| `arrange()` | Change the order of observations (rows)                 |
| `group_by()`| Organize the observations (rows) into groups            |
| `summarise()`| Create summary variables for groups of observations    |


These functions perform actions on single tables of data. Despite their apparent simplicity, combining these operations can accomplish complex data analyses. Hadley Wickham famously stated that 90% of data analysis tasks can be accomplished using these functions. Throughout this chapter, we'll introduce these functions to demonstrate their capabilities.


## Activity 1: Change column names

We are going to learn how to organise data using the *tidy* format^[(http://vita.had.co.nz/papers/tidy-data.pdf)]. This is because we are using the `tidyverse` packages @R-tidyverse. This is an opinionated, but highly effective method for generating reproducible analyses with a wide-range of data manipulation tools. Tidy data is an easy format for computers to read. It is also the required data structure for our **statistical tests** that we will work with later.

Here 'tidy' refers to a specific structure that lets us manipulate and visualise data with ease. In a tidy dataset each *variable* is in one column and each row contains one *observation*. Each cell of the table/spreadsheet contains the *values*. One observation you might make about tidy data is it is quite long - it generates a lot of rows of data - you might remember then that *tidy* data can be referred to as *long*-format data (as opposed to *wide* data). 

```{r, eval=TRUE, echo=FALSE, out.width="80%", fig.alt= "tidy data overview"}
knitr::include_graphics("images/tidy-1.png")
```

So we know our data is in R, and we know the columns and names have been imported. But we still don't know whether all of our values imported correctly, or whether it captured all the rows. 

#### Add this to your script 

```{r, eval = F}
# CHECK DATA----
# check the data
colnames(penguins)
#__________________________----
```

When we run `colnames()` we get the identities of each column in our dataframe

* **Study name**: an identifier for the year in which sets of observations were made

* **Region**: the area in which the observation was recorded

* **Island**: the specific island where the observation was recorded

* **Stage**: Denotes reproductive stage of the penguin

* **Individual** ID: the unique ID of the individual

* **Clutch completion**: if the study nest observed with a full clutch e.g. 2 eggs

* **Date egg**: the date at which the study nest observed with 1 egg

* **Culmen length**: length of the dorsal ridge of the bird's bill (mm)

* **Culmen depth**: depth of the dorsal ridge of the bird's bill (mm)

* **Flipper Length**: length of bird's flipper (mm)

* **Body Mass**: Bird's mass in (g)

* **Sex**: Denotes the sex of the bird

* **Delta 15N** : the ratio of stable Nitrogen isotopes 15N:14N from blood sample

* **Delta 13C**: the ratio of stable Carbon isotopes 13C:12C from blood sample


#### Clean column names

Often we might want to change the names of our variables. They might be non-intuitive, or too long. Our data has a couple of issues:

* Some of the names contain spaces

* Some of the names have capitalised letters

* Some of the names contain brackets

This dataframe  does not like these so let's correct these quickly. R is case-sensitive and also doesn't like spaces or brackets in variable names

```{r, eval = T, warning = F, message = F}
# CLEAN DATA ----

# clean all variable names to snake_case using the clean_names function from the janitor package
# note we are using assign <- to overwrite the old version of penguins with a version that has updated names
# this changes the data in our R workspace but NOT the original csv file

penguins <- janitor::clean_names(penguins) # clean the column names

colnames(penguins) # quickly check the new variable names


```

#### Rename columns (manually)

The `clean_names` function quickly converts all variable names into snake case. The N and C blood isotope ratio names are still quite long though, so let's clean those with `dplyr::rename()` where "new_name" = "old_name".

```{r, eval = T, warning = F, message = F}
# shorten the variable names for N and C isotope blood samples

penguins <- rename(penguins,
         "delta_15n"="delta_15_n_o_oo",  # use rename from the dplyr package
         "delta_13c"="delta_13_c_o_oo")

```


## Activity 2: Look at the data

Have a look at your data by typing the word `penguins` into your console window and running it.

```{r, eval = F}
penguins
```

We can see a dataset with 345 rows (including the headers) and 19 variables
It also provides information on the *type* of data in each column

* `<chr>` - means character or text data

* `<dbl>` - means numerical data

#### Rename text values

Sometimes we may want to rename the values in our variables in order to make a shorthand that is easier to follow. This is changing the **values** in our columns, not the column names. 

```{r, eval = T, warning = F, message = F}
# use mutate and case_when for a statement that conditionally changes the names of the values in a variable
penguins <- penguins %>% 
  mutate(species = case_when(species == "Adelie Penguin (Pygoscelis adeliae)" ~ "Adelie",
                             species == "Gentoo penguin (Pygoscelis papua)" ~ "Gentoo",
                             species == "Chinstrap penguin (Pygoscelis antarctica)" ~ "Chinstrap"))

```


```{block, type = "warning"}

Have you checked that the above code block worked? Inspect your new tibble and check the variables have been renamed as you wanted.

```


## dplyr verbs

Ok so now we have a fair understanding of the penguins data, let's use it to learn a bit more about the six functions from dplyr that make up a lot of data wrangling!

### Activity 3: selecting variables with `select()`

If we wanted to create a dataset that only includes certain variables, we can use the `select()` function from the `dplyr` package. 

For example I might wish to create a simplified dataset that only contains `species`, `sex`, `flipper_length_mm` and `body_mass_g`. 

Run the below code to select only those columns

```{r, eval = F}

# DPLYR VERBS ----

select(.data = penguins, # the data object
       species, sex, flipper_length_mm, body_mass_g) # the variables you want to select

```

Alternatively you could tell R the columns you **don't** want e.g. 

```{r, eval = F}
select(.data = penguins,
       -study_name, -sample_number)

```

Note that `select()` does **not** change the original `penguins` tibble. It spits out the new tibble directly into your console. 

If you don't **save** this new tibble, it won't be stored. If you want to keep it, then you must create a new object. 

When you run this new code, you will not see anything in your console, but you will see a new object appear in your Environment pane.

```{r, eval = F}
new_penguins <- select(.data = penguins, 
       species, sex, flipper_length_mm, body_mass_g)
```

### Activity 4: using `filter()` to choose observations

Having previously used `select()` to select certain variables, we will now use `filter()` to select only certain rows or observations. For example only Adelie penguins. 

We can do this with the equivalence operator `==`

```{r, eval = F}
filter(.data = new_penguins, species == "Adelie")
```

Filter is quite a complicated function, and uses several different operators to assess the way in which it should apply a filter.

```{r, eval = TRUE, echo = FALSE, warning = FALSE, message = FALSE}

Operator <- c("A < B", "A <= B", "A > B", "A >= B", "A == B", "A != B", "A %in% B")

Name <- c("less than", "less than or equal to", "greater than", "greater than or equal to", "equivalence", "not equal", "in")

box <- tibble(Operator, Name)

box %>% 
   kable(caption = "Boolean expressions", 
    booktabs = T) %>% 
   kableExtra::kable_styling(full_width = FALSE, font_size=16)

```

If you wanted to select all the Penguin species except Adelies, you use 'not equals'.

```{r, eval = F}
filter(.data = new_penguins, species != "Adelie")

```

This is the same as 

```{r, eval = F}
filter(.data = new_penguins, species %in% c("Chinstrap", "Gentoo"))
```

You can include multiple expressions within `filter()` and it will pull out only those rows that evaluate to `TRUE` for all of your conditions. 

For example the below code will pull out only those observations of Adelie penguins where flipper length was measured as greater than 190mm. 

```{r, eval = F}
filter(.data = new_penguins, species == "Adelie", flipper_length_mm > 190)
```

### Activity 5: Arranging data with `arrange()`

The function `arrange()` sorts the rows in the table according to the columns supplied. For example

```{r, eval = F}
arrange(.data = new_penguins, sex)
```

The data is now arranged in alphabetical order by sex. So all of the observations of female penguins are listed before males. 

You can also reverse this with `desc()`

```{r, eval = F}

arrange(.data = new_penguins, desc(sex))

```

You can also sort by more than one column, what do you think the code below does?

```{r, eval = F}
arrange(.data = new_penguins,
        sex,
        desc(species),
        desc(flipper_length_mm))
```

### Activity 6: Create new variables with `mutate()`

Sometimes we need to create a new variable that doesn't exist in our dataset. For example we might want to figure out what the flipper length is when factoring in body mass. 

To create new variables we use the function `mutate()`. 

Note that as before, if you want to save your new column you must save it as an object. Here we are mutating a new column and attaching it to the `new_penguins` data oject.

```{r, eval = F}
new_penguins <- mutate(.data = new_penguins,
                       body_mass_kg = body_mass_g/1000)
```



### Activity 7: data summaries with `summarise()`

- `summarise()` is the function we use to create summaries of the data, as per usual the first argument is the object we want to use, in this case `penguins`.

- The second argument is the name of the summary column, in this case `total`, and the summary value we want to create, in this case the total number of rows, using `n()`, we then made a third column of the mean body mass of the penguins (ignoring missing values).

```{r, eval = F}
summarise(
  .data = penguins,     # The dataset to summarize
  total = n(),          # Calculate the total number of rows
  mean = mean(body_mass_g, na.rm = TRUE)  # Calculate the mean of 'body_mass_g', ignoring NA values
)

```

The minimum weight for our penguins is 2.7kg, and the max is 6.3kg - not outrageous. If the min had come out at 27g we might have been suspicious. We will use `summarise` again to calculate other metrics in the future. 

```{block, type = "info"} 
our first data insight, the difference the smallest adult penguin in our dataset is nearly half the size of the largest penguin. 
```

### Activity 8: grouping calculations with `group_by()`

Many data analysis tasks can be approached using the “split-apply-combine” paradigm: split the data into groups, apply some analysis to each group, and then combine the results. `dplyr` makes this very easy with the `group_by()` function. In the `summarise` example above we were able to find the max-min body mass values for the penguins in our dataset. But what if we wanted to break that down by a grouping such as species of penguin. This is where `group_by()` comes in.

```{r, eval = F}

group_penguins <- group_by(penguins, species) # subsequent functions are perform "by group"

summarise(
  .data = group_penguins,     # The dataset to summarize
  total = n(),          # Calculate the total number of rows
  mean = mean(body_mass_g, na.rm = TRUE)  # Calculate the mean of 'body_mass_g', ignoring NA values
)
```

Now we know a little more about our data, the mean weight of our Gentoo penguins is much larger than the other two species. And there are around half as many Chinstrap penguins in our dataset as Gentoo or Adelie. 


## Introducing Pipes

```{r, eval=TRUE, echo=FALSE, out.width="80%", fig.alt= "Pipes make code more human readable"}
knitr::include_graphics("images/pipe_order.jpg")
```

Pipes look like this: `%>%` Pipes allow you to send the output from one function straight into another function. Specifically, they send the result of the function before `%>%` to be the **first** argument of the function after `%>%`. As usual, it's easier to show, rather than tell so let's look at an example.

```{r, eval = F}
# this example uses brackets to nest and order functions
arrange(.data = filter(.data = select(.data = penguins, species, sex, flipper_length_mm), sex == "MALE"), desc(flipper_length_mm))

```

```{r, eval = F}
# this example uses sequential R objects to make the code more readable
object_1 <- select(.data = penguins, species, sex, flipper_length_mm)
object_2 <- filter(.data = object_1, sex == "MALE")
arrange(object_2, desc(flipper_length_mm))

```

```{r, eval = F}
# this example is human readable without intermediate objects
penguins %>% 
  select(species, sex, flipper_length_mm) %>% 
  filter(sex == "MALE") %>% 
  arrange(desc(flipper_length_mm))
```

The reason that this function is called a pipe is because it 'pipes' the data through to the next function. When you wrote the code previously, the first argument of each function was the dataset you wanted to work on. When you use pipes it will automatically take the data from the previous line of code so you don't need to specify it again.

```{task}
Try and write out as plain English what the %>% above is doing? You can read the %>% as THEN
```

`r hide("Solution")`

Take the penguins data AND THEN
Select only the species, sex and flipper length columns AND THEN
Filter to keep only those observations labelled as sex equals male AND THEN
Arrange the data from HIGHEST to LOWEST flipper lengths.

`r unhide()`

```{block, type = "info"}

From R version 4 onwards there is now a "native pipe" `|>`

This doesn't require the tidyverse `magrittr` package or any other packages to load and use. 

For this coursebook I have chosen to continue to use the `tidyverse` pipe `%>%` for the time being it is likely to be much more familiar in other tutorials, and website usages. The native pipe also behaves "slightly" differently, and this could cause some confusion. 

If you want to read about some of the operational differences, [this site](https://www.infoworld.com/article/3621369/use-the-new-r-pipe-built-into-r-41.html) does a good job of explaining 



```


### Activity 9: Using pipes to group and summarise

```{r}

penguins %>% 
  group_by(species) %>% 
  summarise(total = n(),    
  mean = mean(body_mass_g, na.rm = TRUE))
  
```

## A few more handy functions

### Check for duplication

It is very easy when inputting data to make mistakes, copy something in twice for example, or if someone did a lot of copy-pasting to assemble a spreadsheet (yikes!). We can check this pretty quickly

```{r, eval = F}
# check for duplicate rows in the data
penguins %>% 
  duplicated() %>% # produces a list of TRUE/FALSE statements for duplicated or not
  sum() # sums all the TRUE statements

```

```
[1] 0
```
Great! 

### Distinct

We can also look for typos by asking R to produce all of the distinct values in a variable. This is more useful for categorical data, where we expect there to be only a few distinct categories

```{r, eval = F}
penguins %>% 
  distinct(sex)

```

Here if someone had mistyped e.g. 'FMALE' it would be obvious. We could do the same thing (and probably should have before we changed the names) for species. 

### Missing values: NA

There are multiple ways to check for missing values in our data

```{r, eval = F}
# Get a sum of how many observations are missing in our dataframe
penguins %>% 
  is.na() %>% 
  sum()

```

But this doesn't tell us where these are, fortunately the function `summary` does this easily

## `Summary`

```{r, eval = F}
# produce a summary of our data
summary(penguins)
#__________________________----
```

This provides a quick breakdown of the max and min for all numeric variables, as well as a list of how many missing observations there are for each one. As we can see there appear to be two missing observations for measurements in body mass, bill lengths, flipper lengths and several more for blood measures. We don't know for sure without inspecting our data further, *but* it is likely that the two birds are missing multiple measurements, and that several more were measured but didn't have their blood drawn. 

We will leave the NA's alone for now, but it's useful to know how many we have. 

We've now got a clean & tidy dataset, with a handful of first insights into the data. 


## Finished

That was a lot of work! But remember you don't have to remember all of these functions, remember this chapter when you do more data wrangling in the future. Also bookmark the [RStudio Cheatsheets Page](https://www.rstudio.com/resources/cheatsheets/). 

Finally, make sure you have saved the changes made to your script 💾 & make sure your workspace is set **not** to save objects from the environment [*between* sessions](#global-options). 

We want our script to be our record of work and progress, and not to be confused by a cluttered R Environment. 


## Activity: Reorganise this script
Using the link below take the text and copy/paste into a **new** R script and save this as `YYYY_MM_DD_workshop_4_jumbled_script.R` 

All of the correct lines of code, comments and document markers are present, but not in the correct order. Can you unscramble them to produce a sensible output and a clear document outline?


```{r, eval = TRUE, echo = FALSE}
downloadthis::download_link(
  link = "https://raw.githubusercontent.com/UEABIO/data-sci-v1/main/book/files/jumbled_script.R",
  button_label = "Download jumbled R script",
  button_type = "success",
  has_icon = TRUE,
  icon = "fa fa-save",
  self_contained = FALSE
)
```

`r hide("Solution")`

If you want to check your answers (or are just completely stuck) then click here

```{r, eval = TRUE, echo = FALSE}
downloadthis::download_link(
  link = "https://raw.githubusercontent.com/UEABIO/data-sci-v1/main/book/files/unjumbled_script.R",
  button_label = "Unjumbled script",
  button_type = "success",
  has_icon = TRUE,
  icon = "fa fa-save",
  self_contained = FALSE
)
```

`r unhide()`


```{r, echo = F}
save(penguins, file = here("book", "files", "chapter4.RData"))
```
