==> book/18-Complex-models.Rmd <==
# Complex models


```{r, child='_setup.Rmd'}

```



```{r, eval=T, echo=F}
library(tidyverse)
library(GGally)
library(ggridges)
library(performance)
library(patchwork)

colours <- c("cyan", "darkorange", "purple")
```

## Designing a Model

We are introduced to the fruitfly dataset Partridge and Farquhar (1981)^[https://nature.com/articles/294580a0]. From our understanding of sexual selection and reproductive biology in fruit flies, we know there is a well established 'cost' to reproduction in terms of reduced longevity for female fruitflies. The data from this experiment is designed to test whether increased sexual activity affects the lifespan of male fruitflies.

The flies used were an outbred stock, sexual activity was manipulated by supplying males with either new virgin females each day, previously mated females ( Inseminated, so remating rates are lower), or provide no females at all (Control). All groups were otherwise treated identically.

```{r, eval = TRUE, echo = FALSE}
fruitfly <- readRDS(here::here("book", "files", "fruitfly.RDS"))
```


```{r, eval = TRUE, echo = FALSE}
downloadthis::download_link(
  link = "https://raw.githubusercontent.com/Philip-Leftwich/physalia-stats-intro/main/book/files/fruitfly.csv",
  button_label = "Download Fruitfly data as csv",
  button_type = "success",
  has_icon = TRUE,
  icon = "fa fa-save",
  self_contained = FALSE
)
```


* **type**: type of female companion (virgin, inseminated, control(partners = 0))

* **longevity**: lifespan in days

* **thorax**: length of thorax in micrometres (a proxy for body size)

* **sleep**: percentage of the day spent sleeping

## Hypothesis

Before you start any formal analysis you should think clearly about the sensible parameters to test. In this example, we are *most* interested in the effect of sexual activity on longevity. But it is possible that other factors may also affect longevity and we should include these in our model as well, and we should think **hard** about what terms might reasonably be expected to *interact* with sexual activity to affect longevity. 

```{task}
Which terms and interactions do you think we should include in our model?
```

```{solution}
In this exercise I have just asked you to try and think logically about suitable predictors. For a more formal investigation you should support this with evidence where possible

* type - should definitely be included. 

* thorax - the size of the flies could determine longevity. Carreira et al (2009)^[https://www.nature.com/articles/hdy2008117]

* sleep - sleep could easily help determine longevity. Thompson et al (2020)^[https://journals.biologists.com/bio/article/9/9/bio054361/225803/Sleep-length-differences-are-associated-with]

* type:sleep - the amount that sleep (rest) helps promote longevity could change depending on how much activity the fly engages in when awake. Chen et al (2017)^[https://www.nature.com/articles/s41467-017-00087-5#:~:text=In%20this%20study%2C%20we%20show,but%20aroused%20females%20sleep%20more]

Other interactions *could* be included but you should have a strong reason for them. 
```

## Checking the data

You should now import, clean and tidy your data. Making sure it is in tidy format, all variables have useful names, and there are no mistakes, missing data or typos.

Based on the variables you have decided to test you should start with some simple visualisations, to understand the distribution of your data, and investigate visually the relationships you wish to test.

This is a full two-by-two plot of the entire dataset, but you should try and follow this up with some specific plots. 

```{r, eval = TRUE, warning = FALSE, message = FALSE}

GGally::ggpairs(fruitfly)

```

## Activity 1: Think about your data

Think carefully about the plots you should make to investigate the potential differences and relationships you wish to investigate - try and answer the questions first before checking the examples hidden behind dropdowns.

```{task}
Make density distributions for longevity of males across the three treatments.
```


```{solution}
In this first figure - we can investigate whether there is an obvious difference in the longevities of males across the three treatments

``{r, message = FALSE, eval = TRUE, fig.cap = "A density distribution of longevity across the three sexual activity treatments"}

colours <- c("cyan", "darkorange", "purple")

fruitfly %>% 
  ggplot(aes(x = longevity, y = type, fill = type))+
  geom_density_ridges(alpha = 0.5)+
  scale_fill_manual(values = colours)+
  theme_minimal()+
  theme(legend.position = "none")


``

```

**Q** Does it like treatment affects longevity? `r mcq(c(answer = "Yes","No"))`

```{task}
Make a scatterplot of size against longevity.
```


```{solution}
In this first figure - we can investigate whether there is an obvious difference in the longevities of males across the three treatments

``{r, message = FALSE, eval = TRUE, fig.cap = "A scatterplot of longevity against body size (thorax (mm)). No trend line added - often it is a good idea to look at data points without being lead to a conclusion by a line"}


fruitfly %>% 
  ggplot(aes(x = thorax, y = longevity))+
  geom_point()+
  theme_minimal()+
  theme(legend.position = "none")


``

```


**Q** Does it look like size affects longevity? `r mcq(c(answer = "Yes","No"))`


```{task}
Make an interaction plot to see if size interacts with treatment to affect longevity. (Use colour and groups to make differentiate points and lines)
```

```{solution}

``{r, message = FALSE, eval = TRUE, fig.cap = "A scatterplot of thorax against longevity - colours indicate treatment types. This time I have included a line, as it will help determine if I think the slopes are different by group"}

colours <- c("cyan", "darkorange", "purple")

fruitfly %>% 
  ggplot(aes(x=thorax, y = longevity, group = type, colour = type))+
  geom_point( alpha = 0.6)+
  geom_smooth(method = "lm",
            se = FALSE)+
  scale_colour_manual(values = colours)+
  theme_minimal()
``

```

**Q** Does it look like size affects longevity differently between treatment groups? `r mcq(c("Yes", answer = "No"))`

`r hide("Explain this")`

Here it does look as though larger flies have a longer lifespan than smaller flies. But there appears to be little difference in the angle of the slopes between groups. This does not mean we can't test this in our model, but we may decide it is not worth including.

`r unhide()`



We are also interested in the potential effect of sleep on activity, we can construct a scatter plot of sleep against longevity, while including treatment as a covariate.

```{task}
Make an interaction plot to see if sleep interacts with treatment to affect longevity.
```

```{solution}

``{r, message = FALSE, eval = TRUE, fig.cap = "A scatter plot of proportion of time spent sleeping against longevity with a linear model trendline. Points represent individual flies, colours represent treatments."}

fruitfly %>% 
  ggplot(aes(x=sleep, y = longevity, group = type, colour = type))+
  geom_point( alpha = 0.6)+
  geom_smooth(method = "lm",
            se = FALSE)+
  scale_colour_manual(values = colours)+
  theme_minimal()
``

```


In these plots - Are the trendlines moving in the same direction?  `r mcq(c("Yes", answer ="No"))`

`r hide("Explain this")`

Here it does look as though sleep interacts with treatment to affect lifespan. As the slopes of the lines are very different in each group. But in order to know the strength of this association, and if it is significantly different from what we might observe under the null hypothesis, we will have to build a model.

`r unhide()`



## Designing a model

```{block, type = "info"}
When you include an interaction term, the numbers produced from this are how much **more** or **less** the mean estimate is than if you just combined the main effects. 
```


```{multCode = T}

``{r}

flyls1 <- lm(longevity ~ type + thorax + sleep + type:sleep, data = fruitfly)
summary(flyls1)
``

####

``{r}
# a full model
flyls1 <- lm(longevity ~ type + thorax + sleep + type:sleep, data = fruitfly)

flyls1 %>% 
  broom::tidy()
``


```


```{block, type = "info"}
Because we have included an interaction effect the number of terms is quite long and takes more consideration to understand. We can see for the individual estimates that it does not appear that the interaction is having a strong effect (estimate) and this does not appear to be different from a null hypothesis of no interaction effect. But we we should use an *F* test to look at the overall effect to be sure. 
```

```{task}
From the model summary table could you say what the mean longevity of a male with a 0.79mm thorax, that sleeps for 22% of the day and is paired with virgin females would be?
```


```{solution}

``{r, eval = F}
# intercept
coef(flyls1)[1] + 
  
# 1*coefficient for virgin treatment  
coef(flyls1)[3] + 
  
# 0.79 * coefficient for thorax size  
(coef(flyls1)[4]*0.79) + 
  
# 22 * coefficient for sleep  
(coef(flyls1)[5]*22) + 

``

``{r}
# 22 * 1 * coefficient for interaction
(coef(flyls1)[7]*22*1)
``

```


## Model checking & collinearity

Before we start playing with the terms in our model, we should check to see if this is even a good way of fitting and measuring our data. We should check the assumptions of our model are being met.

```{r, eval = TRUE, message = FALSE, warning = FALSE, fig.height=9}

performance::check_model(flyls1, detrend = F)

```

## Activity 2: Model checking

**Question - IS the assumption of homogeneity of variance met?** `r mcq(c( answer = "Yes", "No"))`

```{solution}

* Mostly - the reference line is fairly flat (there is a slight curve).

* It looks as though there might be some increasing heterogeneity with larger values, though very minor.

VERDICT, pretty much ok, should be fine for making inferences. 

With a slight curvature this could indicate that you *might* get a better fit with a transformation, or perhaps that there is a missing variable that if included in the model would improve the residuals. In this instance I wouldn not be overly concerned. See here for a great explainer on intepreting residuals^[https://www.qualtrics.com/support/stats-iq/analyses/regression-guides/interpreting-residual-plots-improve-regression/].


``{r}
performance::check_heteroscedasticity(flyls1)

``

```

**Question - ARE the residuals normally distributed?** `r mcq(c(answer = "Yes", "No"))`

```{solution}
Yes - the QQplot looks pretty good, a very minor indication of a right skew, but nothing to worry about. 

[Interpreting QQ plots][What is a Quantile-Quantile (QQ) plot?]

``{r}
performance::check_normality(flyls1)

``

```


**Question - IS their an issue with Collinearity?** `r mcq(c("Yes", answer ="No"))`

```{solution}

This graph clearly shows there **is** collinearity. But this is not unusual when we include an *interaction term*, if we see evidence of collinearity in terms that are not part of an interaction **then** we should take another look^[https://easystats.github.io/performance/reference/check_collinearity.html].

What can you do about collinearity in main effects? 1) Nothing 2) Mean-center 3) Drop one of the terms. 

The `check_performance()` function produces a visual summary of a Variance Inflation Factor produced from the `vif()` function. This is a measure of the standard error of each estimated coefficient. If this is very larger (greater than 5 or 10), this indicates the model has problems estimating the coefficient. This does not affect model predictions, but makes it more difficult to determine the estimate change from a predictor. 

``{r}
performance::check_collinearity(flyls1)

``

```


## Mean-centering

We often find issues of apparent collinearity are reduced or removed when we mean-center continuous variables, this has the added bonus of centering our intercept, so let's try that now! 

```{r}

fruitfly <- fruitfly %>% 
  mutate(thorax_centered = thorax - mean(thorax),
         sleep_centered = sleep - mean(sleep))

flyls1 <- lm(longevity ~ type + thorax_centered + sleep_centered + type:sleep_centered, data = fruitfly)

performance::check_model(flyls1, check = "vif")

```

## Data transformations

The most common issues when trying to fit simple linear regression models is that our response variable is not normal which violates our modelling assumption. There are two things we can do in this case:

* Variable transformation e.g `lm(sqrt(x) ~ y, data = data)`
    
    - Can sometimes fix linearity
    
    - Can sometimes fix non-normality and heteroscedasticity (i.e non-constant variance) 
    
* Generalized Linear Models (GLMs) to change the error structure (i.e the assumption that residuals need to be normal - see next week.)

### BoxCox

```{block, type="info"}
The BoxCox gets its name from its two inventors, George Box and David Cox. 
Implemented by the MASS package, when applied to a linear model it sytematically applies transformations by raising the y variable to a power (lambda).

The R output for the `MASS::boxcox()` function plots a maximum likelihood curve (with a 95% confidence interval - drops down as dotted lines) for the best transformation for fitting the data to the model.

```

In this case we already know that our model is very likely "good enough" as it does not deviate from expectations of our linear model assumptions - however it is possible that we can still find a better fit. 


```{r, eval = TRUE, echo = FALSE, warning = FALSE, message = FALSE}
library(kableExtra)

`lambda value` <- c(0, 0.5, 1, 2 )

`transformation` <- c("log(Y)", "sqrt(Y)", "Y", "Y^1")

box <- tibble(`lambda value`, `transformation`)

box %>% 
   kbl(caption = "Common Box-Cox Transformations", 
    booktabs = T) %>% 
   kable_styling(full_width = FALSE, font_size=16)

```

```{r, fig.cap = "standard curve fitted by maximum likelihood, dashed lines represent the 95% confidence interval range for picking the 'best' transformation for the dependent variable"}
# run this, pick a transformation and retest the model fit
MASS::boxcox(flyls1)

```

**Question - Does the fit of the model improve with a square root transformation?** `r mcq(c("Yes", answer ="No"))`

```{solution}

``{r}
flyls_sqrt <- lm(sqrt(longevity) ~ type + thorax + sleep + type:sleep, data = fruitfly)

performance::check_model(flyls_sqrt)
``

Not really, despite the suggestion that a sqrt transformation would improve the model, residual fits are not really any better - if we examine R squared carefully, we can see that there is a very small increase in variance explained - but in my opinion, not worth the hassle of having a transformed dependent variable - so we might as well stick with the original scale.




```

## Model selection

```{r}
# use drop1 function to remove top-level terms
drop1(flyls1, test = "F")

```

Based on this ANOVA table, we do not appear to have a strong rationale for keeping the interaction term in the model (AIC or F-test). Therefore we can confidently remove the interaction, simplifying our model and making interpretation easier. 

> Here I have not used the mean-centered variable for thorax - remember this makes no difference to the model except for interpreting the intercept value

```{r, eval = TRUE, echo = F}
flyls2 <- lm(longevity ~ type + thorax + sleep, data = fruitfly)

```


```{r}

flyls2 <- lm(longevity ~ type + thorax + sleep, data = fruitfly)

drop1(flyls2, test = "F")

```

**Question - Should we drop sleep from this model?** `r mcq(c("Yes", answer ="No"))`

```{solution}

There is good reason to remove non-significant *interaction terms* from a model, they complicate estimates and make interpretations more difficult. For **main** effects things are a little more ambiguous. 

When the main aim is prediction, it makes sense to be cautious and retain non-significant terms, as extra terms make no difference to the R^2 of a model. 

When the focus is on hypothesis testing, then removal of non-significant terms can help produce a 'true' model, but this is optional. Generally speaking it is often simpler to leave main effects in the model (you should have carefully considered the terms which were included in the first place). 

In this example we can also see that AIC has not really changed - so the quality of the model is also not improved vby dropping this term. 
```


## Posthoc

Using the [emmeans](https://aosmith.rbind.io/2019/03/25/getting-started-with-emmeans/) package is a very easy way to produce the estimate mean values (rather than mean differences) for different categories `emmeans`. If the term `pairwise` is included then it will also include post-hoc pairwise comparisons between all levels with a tukey test `contrasts`.

```{r}

emmeans::emmeans(flyls2, specs = pairwise ~ type + thorax + sleep)

```

```{block, type = "info"}
For continuous variables (sleep and thorax) - `emmeans` has set these to the mean value within the dataset, so comparisons are constant between categories at the average value of all continuous variables. 
```

If we provide the data ranges, then we can get predictions built from the model across the observed range of predictor variables

```{r, eval = F}

  emmeans::emmeans(flyls2, specs = ~ type + thorax + sleep,
                 at=list(thorax=c(.64:.94),
                         sleep = c(1:83))) %>% 
  as_tibble()

```

## Activity 3: Write-up

```{task}
Can you write an Analysis section?
```

```{solution}
I constructed an ordinary least squares model to investigate the effects of sleep, mating type and body size on longevity in adult Drosophila melanogaster. I also included an interaction term between sleep and mating type. All Analyses and data cleaning was carried out in R ver 4.1.2 with the tidyverse range of packages (Wickham et al 2019), model residuals were checked with the performance package (Lüdecke et al 2021), and summary tables produced with broom (Robinson et al 2022) and kableExtra (Zhu 2020).
```


```{task} 
Can you write a Results section?
```

```{solution}
I tested the hypothesis that sexual activity is costly for male *Drosophila melanogaster* fruitflies. Previous research indicated that sleep deprived males are less attractive to females, this would indicate that levels of sexual activity might be affected by sleep and impact the effect on longevity, as such this was included as an interaction term in the full model. Body size is also know to affect lifespan, as such this was included as a covariate in the mode. 

There was a small interaction effect of decreased lifespan with increasing sleep in the treatment groups compared to control in our samples, but this was not significantly different from no effect (F~2,118~ = 0.512, P = 0.6), and was therefore dropped from the full model (Table 15.1). 

``{r}
library(kableExtra)
flyls2 %>% broom::tidy(conf.int = T) %>% 
 select(-`std.error`) %>% 
mutate_if(is.numeric, round, 2) %>% 
kbl(col.names = c("Predictors",
                    "Estimates",
                    "Z-value",
                    "P",
                    "Lower 95% CI",
                    "Upper 95% CI"),
      caption = "Linear model coefficients", 
    booktabs = T) %>% 
   kable_styling(full_width = FALSE, font_size=16)
``

There was a significant overall effect of treatment on male longevity (Linear model: F~2,120~ = 30.1, P < 0.001), with males paired to virgin females having the lowest mean longevity (48 days, [95%CI: 44.9 - 51.2]) (when holding body size and sleep constant), compared to control males (61.3 days [56.8 - 65.8]) and males paired with inseminated females (64.9 days [61.8 - 68.1 days]). 

Post hoc analysis showed that these differences were statistically significant for males paired with control females compared to the inseminated (Tukey test: t~120~ = 4.8, P < 0.001)  and virgin groups (t~120~ = 7.5, P < 0.001), but there was no overall evidence of a difference between inseminated and virgin groups (t~120~ = -1.309  P < 0.3929) (Figure 19.4). 

Comparing the treatment effects against other predictors of longevity such as body size and sleep, I found that sleep had a very small effect on longevity (mean change -0.05 days [-0.18 - 0.07]) which was not significantly different from no effect (Linear model: F~1,120~ = 0.68, P = 0.41). Body size (taken from thorax length) was a significant predictor of longevity (F~1,120~ = 121, P < 0.001), with each 0.1 mm increase in body size adding 14.4 days to the individual lifespan [11.8 - 17]. It appears as though body size has a stronger effect on longevity than treatment, indicating that while there is a measurable cost of sexual activity to males, it may be less severe than in females (not compared here), and less severe than other measurable predictors. 

``{r,eval = TRUE, echo = F, message = F, warning = F, fig.cap = " A scatter plot of longevity against body size across three treatments of differening male sexual activity. Fitted model slopes are from the reduced linear model (main effects only of thorax size, sleep and treatment group), with 95% confidence intervals, circles are individual data points. Marginal plots are density plot distributions for thorax length and longevity split by treatments."}


# optional extras for adding animal silhouette
#library(rphylopic) rphylopic not working
# search for the right image in the database
#fruitfly_phylo <- name_search(text = "Drosophila melanogaster", options = "namebankID")[[1]] 

# name_images(uuid = fruitfly$uid[1])

# extract the ID of thr right image
#fruitfly_2 <- name_images(uuid = fruitfly_phylo$uid[1])$same[[2]]$ui

# get the actual image
#fruitfly_pic <- image_data(fruitfly_2, size = 256)[[1]]

# density plot of thorax length by treatment
marginal1 <- fruitfly %>% 
  ggplot()+
  geom_density(aes(x = thorax, fill = type),
               alpha = 0.5)+
  scale_fill_manual(values = colours)+
  theme_void()+
  theme(legend.position = "none")

# density plot of longevity by treatment
marginal2 <- fruitfly %>% 
  ggplot()+
  geom_density(aes(x = longevity, fill = type),
               alpha = 0.5)+
  scale_fill_manual(values = colours)+
  theme_void()+
  coord_flip()


# create a new dataset where all sleep values are set to a constant (mean) values - see the emmeans table

fruitfly_2 <- fruitfly %>% 
  mutate(sleep = mean(sleep))

# use the final model to produce model predictions set to the new constant sleep dataframe
model_plot <- broom::augment(flyls2, newdata = fruitfly_2, interval = "confidence") %>% 
  ggplot(aes(x=thorax, y = .fitted, colour = type, fill = type))+
  geom_ribbon(aes(ymin=.lower, ymax=.upper), alpha = 0.2)+
    geom_line(linetype = "dashed", show.legend = FALSE)+
  geom_point(data = fruitfly, aes(x = thorax, y = longevity),
             show.legend = FALSE)+
  scale_colour_manual(values = colours)+
  scale_fill_manual(values = colours)+
  labs(y = "Lifespan in days",
       x = "Thorax length (mm)",
       fill = "Type of female exposure")+
  guides(colour = "none")+
  theme_classic()+
  theme(legend.position = "none")
#  add_phylopic(fruitfly_pic, alpha = 0.8, x = 0.7, y = 80, ysize = 10)



layout <- "
AAA#
BBBC
BBBC
BBBC"


marginal1+model_plot+marginal2 +plot_layout(design = layout)
``

```

## Summary

In this chapter we have worked with our scientific knowledge to develop testable hypotheses and built statistical models to formally assess them. We now have a working pipeline for tackling complex datasets, developing insights and producing and explaining robust linear models. 

### Checklist

* Think carefully about the hypotheses to test, use your scientific knowledge and background reading to support this

* Import, clean and understand your dataset: use data visuals to investigate trends and determine if there is clear support for your hypotheses

* Fit a linear model, including interaction terms with caution

* Investigate the fit of your model, understand that parameters may never be perfect, but that classic patterns in residuals may indicate a poorly fitting model - sometimes this can be fixed with careful consideration of missing variables or through data transformation

* Test the removal of any interaction terms from a model, look at AIC and significance tests

* Make sure you understand the output of a model summary, sense check this against the graphs you have made

* The direction and size of any effects are the priority - produce estimates and uncertainties. Make sure the observations are clear.

* Write-up your significance test results, taking care to report not just significance (and all required parts of a significance test). Do you know *what* to report? Within a complex model - reporting *t* will indicate the slope of the line for that single term against the intercept, *F* is the overall effect of a predictor across all levels, *post-hoc* if you wish to compare across all levels. 

* Well described tables and figures can enhance your results sections - take the time to make sure these are informative and attractive. 

# (PART\*) Summary tables {.unnumbered}

# Summary tables

While we need to report the parameters and statistics of our models in text with any reports, it can also be extremely useful to provide a model summary table, especially when reporting complex analyses. Below I present examples from two of my favourite packages for model reporting.

Both packages come with lots of options for customisation to produce presentation ready tables

## sjPlot

`sjPlot` A really nice package that helps produce model summaries for you automatically, currently it will only output in HTML format

```{r}
library(sjPlot)
tab_model(flyls2)
```


```{r}
tab_model(
  flyls2, 
  pred.labels = c("Intercept", "Inseminated females", "Virgin females", "Thorax size (μm
)", "Sleep (%)"),
  dv.labels = "Male longevity",
  string.pred = "Coefficient",
  string.ci = "Conf. Int (95%)",
  string.p = "P-Value"
)


```

`gtsummary` Is an extension of the `gt` package and provides even greater levels of flexibility and customisation. It is also compatible with HTML, word or pdf. Perhaps the only downside is the level of options and customisability can be slightly overwhelming. In addition the `gt` package has been undergoing a lot of active development, so sometimes older code examples are now defunct.


```{r}
library(gtsummary)
tbl_regression(flyls2)

```

```{r}


tbl_regression(flyls2,
               intercept = TRUE,
               label = list(type ~ "Female mating type", thorax ~ "Thorax", sleep ~ "Sleep"),
               pvalue_fun = ~ style_pvalue(.x, digits = 2)) %>% 
  modify_header(label = "Coefficient",
                estimate = "Estimate") %>% 
  bold_labels() %>% 
  bold_p(t = 0.05) %>% 
  as_gt () %>% 
   gt::tab_source_note(gt::md("*This is an ordinary least squares model*"))

```
